refactor: Sidebar 컴포넌트화 & Map 컴포넌트 제거
