+++ package.json
@@ -17,6 +17,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
+    "react-kakao-maps-sdk": "^1.1.5",
     "react-query": "^3.39.2",
     "react-router-dom": "^6.7.0",
     "react-scripts": "5.0.1",
