+++ public/index.html
@@ -27,11 +27,11 @@
       work correctly both with client-side routing and a non-root public URL.
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
+    <title>WHATNYANG</title>
     <script
       type="text/javascript"
       src="//dapi.kakao.com/v2/maps/sdk.js?appkey=234cedba4ef5f78f63f579672c1e7789&libraries=services"
     ></script>
-    <title>WHATNYANG</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
